% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleValueChain.R
\docType{data}
\name{ExampleValueChain}
\alias{ExampleValueChain}
\title{Example of an 'Automotive Value Chain'.}
\format{
A data frame with 175 rows and 3 variables:
\describe{
  \item{tierNumber}{Tier number of product}
  \item{GVCactivityNumber}{Activity number of product}
  \item{HScode}{Exporter, HS17 product code}
}
}
\source{
\url{https://www.unido.org/sites/default/files/2009-12/Global_automotive_industry_value_chain_0.pdf}
}
\usage{
ExampleValueChain
}
\description{
A data set containing the categories for the automotive industry.
}
\keyword{datasets}
