% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iopspackage2.0.R
\name{IOPS}
\alias{IOPS}
\title{IOPS}
\usage{
IOPS(
  CountryCode,
  tradeData,
  ComplexMethod = "eigenvalues",
  iterCompl = 20,
  GVCMapping,
  tradedigit = 6
)
}
\arguments{
\item{CountryCode}{(Type: character/integer) Any accepted ISO country code could be used, e.g. \code{"United Kingdom"}, \code{"GBR"}, \code{"GB"}, \code{828} would all be accepted if the United Kingdom is the desired country. This version of the package uses \code{country\_codes\_V202201} from the BACI Hs17 2020 dataset, but it can be changed in \code{inst/extdata}.}

\item{tradeData}{(Type: csv) Accepts any CEPII BACI trade data. NOTE: \code{tradeData} and \code{GVCMapping} must be from the same "H" Family, e.g., both are from H3, etc., for the program to work correctly.}

\item{ComplexMethod}{(Type: character) Methods used to calculate complexity measures. It can be any one of these methods: \code{"fitness"}, \code{"reflections"}, or \code{"eigenvalues"}. Defaults to "eigenvalues".}

\item{iterCompl}{(Type: integer) The number of iterations that the chosen complexity measure must use. Defaults to \code{iterCompl = 20}.}

\item{GVCMapping}{(Type: csv) The desired value chain to be analyzed. With columns "Tiers," "Activity," and "HSCode." NOTE: \code{tradeData} and \code{GVCMapping} must be from the same "H" Family, e.g., both are from H3, etc., for the program to work correctly.}

\item{tradedigit}{(Type: integer) Indicate if the raw trade digit summation should be done on a 6- or 4-digit level. Defaults to \code{tradedigit = 6}.}
}
\value{
A list that contains the following components:

ECI: (Type: numeric) The Economic Complexity Index.

PCI: (Type: numeric) The Product Complexity Index.

\code{Opportunity\_Gain}: (Type: numeric) Opportunity gain.

distance: (Type: numeric) Distance metric.

density: (Type: numeric) Density metric.

\code{M\_absolute}: (Type: numeric) Absolute metric.

\code{M\_binary}: (Type: numeric) Binary metric.

\code{Tier\_Results}: (Type: data.frame) Results at the tier level.

\code{Product\_Category\_Results}: (Type: data.frame) Results at the product category level.

\code{Product\_Results}: (Type: data.frame) Results at the product level.
}
\description{
Takes user-inputted trade data, any accepted ISO country code, and industrial value chain mapping to calculate various metrics (Economic- and Product complexity metrics, distance metrics, opportunity gain, and inequality metrics) of a given country to facilitate better decision-making regarding industrial policymaking.
}
\examples{
IOPS(
  CountryCode = 710,
  tradeData = ExampleTradeData,
  ComplexMethod = "fitness",
  iterCompl = 20,
  GVCMapping = ExampleValueChain,
  tradedigit = 6
)
}
